Rename Benchmark classes for consistency
Benchmarks in `presto-benchmark` still do not follow the `Benchmark*`
pattern.
