+++ presto-main/src/test/java/io/prestosql/sql/gen/BenchmarkInCodeGenerator.java
@@ -64,7 +64,7 @@ import static org.openjdk.jmh.annotations.Mode.AverageTime;
 @Measurement(iterations = 10, time = 500, timeUnit = TimeUnit.MILLISECONDS)
 @BenchmarkMode(AverageTime)
 @SuppressWarnings({"FieldMayBeFinal", "FieldCanBeLocal"})
-public class InCodeGeneratorBenchmark
+public class BenchmarkInCodeGenerator
 {
     @Param({"1", "5", "10", "25", "50", "75", "100", "150", "200", "250", "300", "350", "400", "450", "500", "750", "1000", "10000"})
     private int inListCount = 1;
@@ -147,7 +147,7 @@ public class InCodeGeneratorBenchmark
     {
         Options options = new OptionsBuilder()
                 .verbosity(VerboseMode.NORMAL)
-                .include(".*" + InCodeGeneratorBenchmark.class.getSimpleName() + ".*")
+                .include(".*" + BenchmarkInCodeGenerator.class.getSimpleName() + ".*")
                 .build();
 
         new Runner(options).run();
